# gin-backend/docs/swagger.yaml
basePath: /
definitions:
  core.ErrorDetail:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  core.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/core.ErrorDetail'
    type: object
  domain.Colonia:
    properties:
      colonia_id:
        type: integer
      created_at:
        type: string
      nombre:
        type: string
      zona:
        type: string
    type: object
  domain.ColoniaListResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/domain.Colonia'
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  domain.ColoniaMessageResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      success:
        type: boolean
    type: object
  domain.ColoniaResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/domain.Colonia'
      message:
        type: string
      success:
        type: boolean
    type: object
  domain.CreateColoniaRequest:
    properties:
      nombre:
        example: Centro Histórico
        type: string
      zona:
        example: Centro
        type: string
    required:
    - nombre
    - zona
    type: object
...
  domain.UpdateColoniaRequest:
    properties:
      nombre:
        example: Centro Histórico
        type: string
      zona:
        example: Centro
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API para gestión de recolección de residuos
  title: API Recolecta
  version: "1.0"
paths:
...
  /api/colonia:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ColoniaListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Listar colonias
      tags:
      - Colonia
    post:
      consumes:
      - application/json
      description: 'Requiere rol Administrador. Body: nombre y zona.'
      parameters:
      - description: Datos de colonia
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateColoniaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.ColoniaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Crear colonia
      tags:
      - Colonia
  /api/colonia/{id}:
    delete:
      parameters:
      - description: ID de colonia
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Eliminar colonia
      tags:
      - Colonia
    get:
      parameters:
      - description: ID de colonia
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ColoniaResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Obtener colonia por ID
      tags:
      - Colonia
    put:
      consumes:
      - application/json
      parameters:
      - description: ID de colonia
        in: path
        name: id
        required: true
        type: integer
      - description: Datos de colonia
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateColoniaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ColoniaMessageResponse'
      security:
      - BearerAuth: []
      summary: Actualizar colonia
      tags:
      - Colonia
...
  /colonias:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Listar colonias (compat)
      tags:
      - Colonia
    post:
      consumes:
      - application/json
      parameters:
      - description: Payload JSON
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateColoniaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Crear colonia (compat)
      tags:
      - Colonia
  /colonias/{id}:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Eliminar colonia (compat)
      tags:
      - Colonia
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Obtener colonia (compat)
      tags:
      - Colonia
    put:
      consumes:
      - application/json
      parameters:
      - description: Payload JSON
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateColoniaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Actualizar colonia (compat)
      tags:
      - Colonia
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Ingresa el token JWT con el prefijo Bearer. Ejemplo: Bearer eyJhbGciOiJIUzI1NiIs...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"